requirements.txt
google-generativeai
pyTelegramBotAPI
flask
requests